move create schedule and delete old API (#3377)
* move create schedule and delete old API

* fix test multitensor
