+++ test/test_conv_shapetracker.py
@@ -3,6 +3,7 @@ import unittest
 from tinygrad.tensor import Tensor
 from tinygrad.ops import LoadOps
 from tinygrad.nn import Conv2d
+from tinygrad.realize import create_schedule
 
 class TestConvShapetracker(unittest.TestCase):
   def test_conv_3x3_one_view(self):
@@ -10,9 +11,9 @@ class TestConvShapetracker(unittest.TestCase):
     seen = set()
 
     # first run to init the weights, they are saved in seen
-    conv(Tensor.empty(1, 16, 10, 10)).lazydata.schedule(seen)
+    create_schedule([conv(Tensor.empty(1, 16, 10, 10)).lazydata], seen)
     # run it again to get the kernels
-    sched = [si for si in conv(Tensor.empty(1, 16, 10, 10)).lazydata.schedule(seen) if si.ast.op not in LoadOps]
+    sched = [si for si in create_schedule([conv(Tensor.empty(1, 16, 10, 10)).lazydata], seen) if si.ast.op not in LoadOps]
     assert len(sched) == 1, f"conv should only have one kernel, getting {len(sched)}"
     print(sched[0])
     for arg in [sched[0].out, *sched[0].inputs]:
