+++ test/test_fusion_op.py
@@ -3,8 +3,7 @@ import time
 import numpy as np
 from tinygrad import Tensor, dtypes
 from tinygrad.device import InterpretedASTRunner
-from tinygrad.lazy import create_schedule
-from tinygrad.realize import run_schedule, lower_schedule_item
+from tinygrad.realize import run_schedule, create_schedule, lower_schedule_item
 
 class TestFusionOp(unittest.TestCase):
   def test_contiguous_add(self):
