+++ test/test_linearizer.py
@@ -10,7 +10,7 @@ from tinygrad.shape.view import View
 from tinygrad.shape.symbolic import MulNode, SumNode, Variable, NumNode, Node, create_rednode
 from tinygrad.tensor import Tensor
 from tinygrad.features.jit import CacheCollector
-from tinygrad.realize import run_schedule
+from tinygrad.realize import create_schedule, run_schedule
 from tinygrad.helpers import prod, Context
 from tinygrad.dtype import DType, dtypes
 
@@ -33,7 +33,7 @@ class TestLinearizer(unittest.TestCase):
     # these are of size 3 to avoid float4 coalesce
     r = a[:-1] + a[1:]
 
-    k = Linearizer(r.lazydata.schedule()[-1].ast)
+    k = Linearizer(create_schedule([r.lazydata])[-1].ast)
     k.upcast()
     k.linearize()
     num_loads = len([uop for uop in k.uops if uop.uop == UOps.LOAD])
@@ -46,7 +46,7 @@ class TestLinearizer(unittest.TestCase):
     a, b = Tensor.randn(1).realize(), Tensor.randn(1).realize()
     r = a.expand([2]) + b.expand([2])
 
-    k = Linearizer(r.lazydata.schedule()[-1].ast)
+    k = Linearizer(create_schedule([r.lazydata])[-1].ast)
     k.upcast()
     k.linearize()
     num_ops = len([uop for uop in k.uops if uop.uop == UOps.ALU])
@@ -56,7 +56,7 @@ class TestLinearizer(unittest.TestCase):
     a, b = Tensor.randn(1).realize(), Tensor.randn(1).realize()
     r = Tensor.stack([a, b])
 
-    k = Linearizer(r.lazydata.schedule()[-1].ast)
+    k = Linearizer(create_schedule([r.lazydata])[-1].ast)
     k.upcast()
     k.linearize()
     num_ops = len([uop for uop in k.uops if uop.uop == UOps.ALU])
@@ -67,7 +67,7 @@ class TestLinearizer(unittest.TestCase):
     a, b = Tensor(2), Tensor(3)
     r = a * b
 
-    k = Linearizer(r.lazydata.schedule()[-1].ast)
+    k = Linearizer(create_schedule([r.lazydata])[-1].ast)
     k.linearize()
     num_ops = len([uop for uop in k.uops if uop.uop in [UOps.LOAD, UOps.ALU]])
     assert num_ops <= 0, "more load or alu uops than needed"
@@ -76,14 +76,14 @@ class TestLinearizer(unittest.TestCase):
     for tensor_dtype, acc_dtype in (
       (dtypes.bool, dtypes.int), (dtypes.int16, dtypes.int), (dtypes.float16, dtypes.float), (dtypes.bfloat16, dtypes.float)):
       a = Tensor([1, 2, 3], dtype=tensor_dtype).sum()
-      k = Linearizer(a.lazydata.schedule()[-1].ast)
+      k = Linearizer(create_schedule([a.lazydata])[-1].ast)
       k.linearize()
       local = [uop for uop in k.uops if uop.uop == UOps.DEFINE_ACC]
       assert local[0].dtype == acc_dtype
 
   def test_arg_acc_dtype(self):
     def helper_arg_acc_dtype(c: Tensor, expected_dtype:DType):
-      k = Linearizer(c.lazydata.schedule()[-1].ast)
+      k = Linearizer(create_schedule([c.lazydata])[-1].ast)
       k.linearize()
       local = [uop for uop in k.uops if uop.uop == UOps.DEFINE_ACC]
       assert local[0].dtype == expected_dtype
@@ -121,7 +121,7 @@ class TestLinearizer(unittest.TestCase):
 
   def test_limit_dims_to_max_5d_global(self):
     t = Tensor.rand(3, 4, 5, 6, 7).pad(((1, 1), (1, 1), (1, 1), (1, 1), (1, 1))) + 1
-    sched = [si for si in t.lazydata.schedule() if si.ast.op not in LoadOps]
+    sched = [si for si in create_schedule([t.lazydata]) if si.ast.op not in LoadOps]
     assert len(sched) == 1
     lin = Linearizer(sched[0].ast)
     assert lin.full_shape[:lin.global_dims] == (5, 6, 7, 8, 9)
@@ -129,7 +129,7 @@ class TestLinearizer(unittest.TestCase):
 
   def test_sum_collapse(self):
     t = Tensor.ones(256,256).sum()
-    sched = [si for si in t.lazydata.schedule() if si.ast.op not in LoadOps]
+    sched = [si for si in create_schedule([t.lazydata]) if si.ast.op not in LoadOps]
     assert len(sched) == 1
     lin = Linearizer(sched[0].ast)
     assert not any(u.uop == UOps.LOOP for u in lin.linearize().uops), "found loop in sum collapse"
@@ -154,7 +154,7 @@ class TestLinearizer(unittest.TestCase):
                                 arg=TernaryOps.WHERE).uop == UOps.ALU
 
 def helper_realized_ast(r:Tensor):
-  s = r.lazydata.schedule()
+  s = create_schedule([r.lazydata])
   run_schedule(s[:-1])  # run all kernels except the last one
   # now all input LazyBuffers buffers in s[-1] should be realized
   # allocate an output buffer
@@ -176,7 +176,7 @@ class TestFloat4(unittest.TestCase):
     b = Tensor.rand(2, 8).realize()
     c = a + b
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()
     k.linearize()
@@ -188,7 +188,7 @@ class TestFloat4(unittest.TestCase):
     b = Tensor.rand(2, 8).realize()
     c = a + b
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.shift_to(0, 4)  # float4 dimension
     k.shift_to(0, 2, insert_before=k.shape_len-1)
@@ -204,7 +204,7 @@ class TestFloat4(unittest.TestCase):
     b = Tensor.rand(9).realize().shrink(((1, 9),))
     c = a + b
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()  # implicit trigger float4 dim
     k.linearize()
@@ -216,7 +216,7 @@ class TestFloat4(unittest.TestCase):
     b = Tensor.rand(2, 9).realize().shrink(((0, 2), (1, 9),))
     c = a + b
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.shift_to(len(k.full_unupcasted_shape)-1, 4)  # manual trigger float4 dim
     k.upcast()
@@ -234,7 +234,7 @@ class TestFloat4(unittest.TestCase):
     # only the first and last conv dot products are aligned in a, and b is never aligned, so no
     # float4 should be emitted (the reduce axis of size 4 is the float4 axis here)
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.upcast()
     k.linearize()
@@ -249,7 +249,7 @@ class TestFloat4(unittest.TestCase):
     # dimension, then we could do float4 for only that one set of loads, but we currently
     # don't.
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.upcast()
     k.upcast()
@@ -265,7 +265,7 @@ class TestFloat4(unittest.TestCase):
     # we will upcast the top axis of sz 4. they should not be coalesced into float4,
     # since the top axis is not contiguous.
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.shift_to(0, 4, top=True)  # top axes are float4 axes
     k.upcast()
@@ -281,7 +281,7 @@ class TestFloat4(unittest.TestCase):
     # we will upcast the top axis of sz 4. they should not be coalesced into float4,
     # since the top axis is not contiguous.
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.shift_to(0, 4)  # float4 axis
     k.upcast()
@@ -296,7 +296,7 @@ class TestFloat4(unittest.TestCase):
 
     # should float4 b but not a
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.shift_to(0, 4)  # float4 axis
     k.upcast()
@@ -310,7 +310,7 @@ class TestHandCodedOpts(unittest.TestCase):
     layer_1 = Tensor.cat(*[Tensor.rand(5) for _ in range(4)])
     layer_2 = Tensor.cat(layer_1.unsqueeze(0), Tensor.rand(6, 20))
 
-    s = layer_2.lazydata.schedule()[-1]
+    s = create_schedule([layer_2.lazydata])[-1]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()
     assert len(k.bufs) == 6  # make sure all ops are done in one kernel
@@ -323,7 +323,7 @@ class TestHandCodedOpts(unittest.TestCase):
   def test_masked_upcast_wino(self):
     monster = Tensor.stack([Tensor.stack([Tensor.rand(16) for _ in range(6)]) for _ in range(6)])
 
-    s = monster.lazydata.schedule()[-1]
+    s = create_schedule([monster.lazydata])[-1]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()
     assert len(k.bufs) == 37  # make sure all ops are done in one kernel
@@ -335,7 +335,7 @@ class TestHandCodedOpts(unittest.TestCase):
       x,w = Tensor.rand(1,4,8,8, requires_grad=True).realize(), Tensor.rand(4,4,3,3, requires_grad=True).realize()
       out = Tensor.conv2d(x,w, padding=1)
       upcasts = []
-      wino_schedule = out.lazydata.schedule()
+      wino_schedule = create_schedule([out.lazydata])
       # collect upcasts of tile transform kernels
       for i, si in enumerate(wino_schedule):
         k = Linearizer(si.ast)
@@ -349,7 +349,7 @@ class TestHandCodedOpts(unittest.TestCase):
       assert upcasts.count((6, 6)) == 2 #and upcasts.count((4, 4)) == 1
 
       out.mean().backward()
-      backward_schedule = x.grad.lazydata.schedule() + w.grad.lazydata.schedule()
+      backward_schedule = create_schedule([x.grad.lazydata, w.grad.lazydata])
       for si in backward_schedule:
         k = Linearizer(si.ast)
         k.hand_coded_optimizations()
@@ -364,7 +364,7 @@ class TestHandCodedOpts(unittest.TestCase):
     layer_2 = Tensor.cat(layer_1.unsqueeze(0), Tensor.rand(6, 7, 4))
     layer_3 = Tensor.cat(layer_2.unsqueeze(0), Tensor.rand(6, 7, 7, 4))
 
-    s = layer_3.lazydata.schedule()[-1]
+    s = create_schedule([layer_3.lazydata])[-1]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()
     assert len(k.bufs) == 5  # make sure all ops are done in one kernel
@@ -379,7 +379,7 @@ class TestHandCodedOpts(unittest.TestCase):
     b = Tensor.rand(N, N).realize()
     c = a @ b
 
-    s = c.lazydata.schedule()[0]
+    s = create_schedule([c.lazydata])[0]
     k = Linearizer(s.ast)
     k.hand_coded_optimizations()
