+++ test/test_schedule.py
@@ -10,16 +10,17 @@ from tinygrad.device import Device, Compiled
 from tinygrad.helpers import DEBUG, GRAPH
 from tinygrad.codegen.linearizer import Linearizer
 from tinygrad.features.graph import print_tree, realized_lazybuffer
+from tinygrad.realize import create_schedule
 from tinygrad import nn, dtypes
 
 def check_schedule(t:Tensor, allowed:int, to_prerealize:Optional[List[Tensor]]=None, filter_loadops=True):
   seen = set()
   if to_prerealize:
     for pre in to_prerealize:
-      for s in pre.lazydata.schedule(seen.copy()):
+      for s in create_schedule([pre.lazydata], seen.copy()):
         if GRAPH: realized_lazybuffer(s.out, 0)
         seen.add(s.out)
-  sched = t.lazydata.schedule(seen)
+  sched = create_schedule([t.lazydata], seen)
   if GRAPH:
     for i,s in enumerate(sched): realized_lazybuffer(s.out, i+1)
   if filter_loadops: sched = [s for s in sched if s.ast.op not in LoadOps]
