+++ test/test_search.py
@@ -1,6 +1,7 @@
 import unittest
 
 from tinygrad.codegen.linearizer import Linearizer
+from tinygrad.realize import create_schedule
 from tinygrad.features.search import time_linearizer
 from tinygrad.device import Compiled, Device, Buffer
 from tinygrad.ops import LoadOps
@@ -11,7 +12,7 @@ class TestTimeLinearizer(unittest.TestCase):
     if not isinstance(Device[Device.DEFAULT], Compiled): raise unittest.SkipTest("only test for compiled backends")
 
   def test_reasonable_time(self):
-    si = [si for si in Tensor([1,2,3,4]).add(1).lazydata.schedule() if si.ast.op not in LoadOps][0]
+    si = [si for si in create_schedule([Tensor([1,2,3,4]).add(1).lazydata]) if si.ast.op not in LoadOps][0]
     rawbufs = [Buffer(Device.DEFAULT, si.out.st.real_size(), si.out.dtype)] + [Buffer(Device.DEFAULT, x.st.real_size(), x.dtype) for x in si.inputs]
     tm = time_linearizer(Linearizer(si.ast), rawbufs, allow_test_size=False, cnt=10)
     assert tm > 0 and tm != float('inf')
